this passes 3e with flying colors
https://fly.io/dist-sys/3e/

but I don't ... I'm not happy with it - it's a lot of messages - in theory I can have all of them randomly send info about current state of everyone instead of sending all updates to everyone on a timer.

1) keep a log for each node and yourself
2) send the log lengh for all nodes randomly to each node every 1s or something as a sync_status message
3) if you get a sync_status that is longer than yours for a node, ask the source node for updates
3a) if you get more updates than the sync status had sync to whoever needs it
4) if you get a sync status that's lower than yours for a node, sync missing data to it

Inspried by SSB but I haven't read anything about it in the last few years.